requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.4.2
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.2.2
django-crispy-forms==1.7.0
django-phonenumber-field==5.1.0
django-widget-tweaks==1.4.8
dnspython==1.16.0
eventlet==0.31.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.0
idna==2.10
isort==5.6.4
jws==0.1.3
kombu==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.1
oauth2client==4.1.3
pandas==1.2.3
phonenumbers==8.12.23
phonenumberslite==8.12.23
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pylint==2.6.0
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
